requirements.txt
streamlit
mcp
google-genai
python-dotenv
requests
contextual-client
